Remove documentation Processor
I revised the whole processor routine. `Processor` used only by `Documenter`
and both `Processor` and `Documenter` share 1/3 of dependencies. With this
in mind, `Processor` basically used to hide one collaborator and attached
behaviour from the `Documenter`. Didn't worth it.

Another interesting this about this commit is that it illustrates why
exactly crossing context boundaries (`Documentation` and `Package` in
this case) is bad. If we'd depend on `Processor` outside of
`Documentation` module, we'd had much harder time making this particular
refactoring.