Refactor DocumentationContributorContext
`:package version :version was documented in :repository` step
definition did diverge from the domain language. The solution was to
either rephrase the step or to rework the step definition, I think the
step itself was fine, so I proceeded with refactoring the Fake infra to
support better language of the context. Result is much more clear and
understandable step definition.